Split dependency solver from checker
