+++ mod/src/test/java/org/shsts/tinactory/gametest/dependency/DependencyChecker.java
@@ -47,20 +47,15 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.NavigableMap;
 import java.util.Optional;
-import java.util.PriorityQueue;
-import java.util.Queue;
 import java.util.Set;
-import java.util.TreeMap;
 import java.util.TreeSet;
 
 import static org.shsts.tinactory.Tinactory.CORE;
 
-public final class DependencyChecker implements IDependencyChecker {
+public final class DependencyChecker {
     private static final ResourceLocation LARGE_CHEMICAL_REACTOR = new ResourceLocation(
         TinactoryKeys.ID, "large_chemical_reactor");
     private static final ResourceLocation MINECRAFT_SMELTING = new ResourceLocation("minecraft", "smelting");
@@ -85,17 +80,7 @@ public final class DependencyChecker implements IDependencyChecker {
     private static final int ACCEPTED_UNREACHABLE_NODES = 41;
 
     private final List<DependencyMethod> methods = new ArrayList<>();
-    private final Queue<DependencyMethod> readyMethods = new PriorityQueue<>();
-    private final Queue<IDependencyNode> newlyReached = new PriorityQueue<>();
-    private final Set<DependencyMethod> appliedMethods = new HashSet<>();
-    private final Map<DependencyMethod, Integer> remainingRequirements = new HashMap<>();
-    private final Map<IDependencyNode, Set<DependencyMethod>> waitingByExactNode = new HashMap<>();
-    private final Map<String, NavigableMap<Double, Set<DependencyMethod>>> waitingByNumericNode = new HashMap<>();
-    private final NavigableMap<Integer, Set<DependencyMethod>> waitingByVoltage = new TreeMap<>();
-    private final Set<IDependencyNode> reachedExactNodes = new TreeSet<>();
-    private final Map<String, Double> maxNumericNodes = new HashMap<>();
     private final Map<IDependencyNode, Set<DependencyMethod>> methodsByOutput = new HashMap<>();
-    private int maxReachedVoltageRank = -1;
 
     public void addMethod(DependencyMethod method) {
         methods.add(method);
@@ -104,169 +89,12 @@ public final class DependencyChecker implements IDependencyChecker {
         }
     }
 
-    public void solve(Collection<IDependencyNode> startNodes) {
-        resetSolver();
-        for (var node : startNodes) {
-            node.reach(this, bootstrapMethod());
-        }
-        initializeMethods();
-        while (!readyMethods.isEmpty() || !newlyReached.isEmpty()) {
-            applyReadyMethods();
-            releaseReachedNodes();
-        }
-    }
-
     public static void runRuntimeCheck(ServerLevel world) {
         var checker = new DependencyChecker();
         checker.extractRuntimeMethods(world);
-        checker.solve(checker.startNodes());
-        checker.writeReport(checker.intendedTargets(), checker.exemptTargets());
-    }
-
-    @Override
-    public boolean isExactReached(IDependencyNode node) {
-        return reachedExactNodes.contains(node);
-    }
-
-    @Override
-    public boolean reachExact(IDependencyNode node, DependencyMethod method) {
-        if (!reachedExactNodes.add(node)) {
-            return false;
-        }
-        newlyReached.add(node);
-        return true;
-    }
-
-    @Override
-    public void addExactWaiter(IDependencyNode node, DependencyMethod method) {
-        waitingByExactNode.computeIfAbsent(node, $ -> new TreeSet<>()).add(method);
-    }
-
-    @Override
-    public Set<DependencyMethod> releaseExactWaiters(IDependencyNode node) {
-        return waitingByExactNode.getOrDefault(node, Set.of());
-    }
-
-    @Override
-    public double maxNumeric(String key) {
-        return maxNumericNodes.getOrDefault(key, Double.NEGATIVE_INFINITY);
-    }
-
-    @Override
-    public boolean reachNumeric(String key, double value, DependencyMethod method) {
-        var oldValue = maxNumeric(key);
-        if (oldValue >= value) {
-            return false;
-        }
-        maxNumericNodes.put(key, value);
-        newlyReached.add(new NumericNode(key, value));
-        return true;
-    }
-
-    @Override
-    public void addNumericWaiter(NumericNode node, DependencyMethod method) {
-        waitingByNumericNode.computeIfAbsent(node.key(), $ -> new TreeMap<>())
-            .computeIfAbsent(node.value(), $ -> new TreeSet<>())
-            .add(method);
-    }
-
-    @Override
-    public Set<DependencyMethod> releaseNumericWaiters(NumericNode node) {
-        var methodsByValue = waitingByNumericNode.get(node.key());
-        if (methodsByValue == null) {
-            return Set.of();
-        }
-        var released = new TreeSet<DependencyMethod>();
-        var reachedValues = new TreeMap<>(methodsByValue.headMap(node.value(), true));
-        for (var methodSet : reachedValues.values()) {
-            released.addAll(methodSet);
-        }
-        reachedValues.keySet().forEach(methodsByValue::remove);
-        return released;
-    }
-
-    @Override
-    public int maxVoltageRank() {
-        return maxReachedVoltageRank;
-    }
-
-    @Override
-    public boolean reachVoltage(Voltage voltage, DependencyMethod method) {
-        if (maxReachedVoltageRank >= voltage.rank) {
-            return false;
-        }
-        maxReachedVoltageRank = voltage.rank;
-        newlyReached.add(new VoltageNode(voltage));
-        return true;
-    }
-
-    @Override
-    public void addVoltageWaiter(VoltageNode node, DependencyMethod method) {
-        waitingByVoltage.computeIfAbsent(node.voltage().rank, $ -> new TreeSet<>()).add(method);
-    }
-
-    @Override
-    public Set<DependencyMethod> releaseVoltageWaiters(VoltageNode node) {
-        var released = new TreeSet<DependencyMethod>();
-        var reachedRanks = new TreeMap<>(waitingByVoltage.headMap(node.voltage().rank, true));
-        for (var methodSet : reachedRanks.values()) {
-            released.addAll(methodSet);
-        }
-        reachedRanks.keySet().forEach(waitingByVoltage::remove);
-        return released;
-    }
-
-    private void resetSolver() {
-        readyMethods.clear();
-        newlyReached.clear();
-        appliedMethods.clear();
-        remainingRequirements.clear();
-        waitingByExactNode.clear();
-        waitingByNumericNode.clear();
-        waitingByVoltage.clear();
-        reachedExactNodes.clear();
-        maxNumericNodes.clear();
-        maxReachedVoltageRank = -1;
-    }
-
-    private void initializeMethods() {
-        for (var method : new TreeSet<>(methods)) {
-            var remaining = 0;
-            for (var requirement : method.requirements()) {
-                if (!requirement.isSatisfied(this)) {
-                    remaining++;
-                    requirement.addWaiter(this, method);
-                }
-            }
-            remainingRequirements.put(method, remaining);
-            if (remaining == 0) {
-                readyMethods.add(method);
-            }
-        }
-    }
-
-    private void applyReadyMethods() {
-        while (!readyMethods.isEmpty()) {
-            var method = readyMethods.remove();
-            if (!appliedMethods.add(method)) {
-                continue;
-            }
-            for (var output : method.outputs()) {
-                output.reach(this, method);
-            }
-        }
-    }
-
-    private void releaseReachedNodes() {
-        while (!newlyReached.isEmpty()) {
-            var node = newlyReached.remove();
-            for (var method : node.releaseWaiters(this)) {
-                var remaining = remainingRequirements.merge(method, -1, Integer::sum);
-                if (remaining == 0) {
-                    readyMethods.add(method);
-                }
-            }
-        }
+        var solver = new DependencySolver(checker.methods);
+        solver.solve(checker.startNodes());
+        checker.writeReport(solver, checker.intendedTargets(), checker.exemptTargets());
     }
 
     private void extractRuntimeMethods(ServerLevel world) {
@@ -653,7 +481,8 @@ public final class DependencyChecker implements IDependencyChecker {
         }
     }
 
-    private void writeReport(Set<IDependencyNode> targets, Map<IDependencyNode, String> exemptions) {
+    private void writeReport(IDependencySolver solver, Set<IDependencyNode> targets,
+        Map<IDependencyNode, String> exemptions) {
         var reportFile = System.getProperty("tinactory.dependencyChecker.reportFile", "");
         if (reportFile.isBlank()) {
             return;
@@ -661,9 +490,9 @@ public final class DependencyChecker implements IDependencyChecker {
         var lines = new ArrayList<String>();
         var missingTargets = 0;
         for (var target : targets) {
-            if (!target.isSatisfied(this) && !exemptions.containsKey(target)) {
+            if (!target.isSatisfied(solver) && !exemptions.containsKey(target)) {
                 missingTargets++;
-                lines.addAll(formatMissingTarget(target));
+                lines.addAll(formatMissingTarget(solver, target));
             }
         }
         var path = Path.of(reportFile);
@@ -687,14 +516,14 @@ public final class DependencyChecker implements IDependencyChecker {
         }
     }
 
-    private Collection<String> formatMissingTarget(IDependencyNode target) {
+    private Collection<String> formatMissingTarget(IDependencySolver solver, IDependencyNode target) {
         var lines = new ArrayList<String>();
         lines.add("missing: " + target.displayId());
         if (Boolean.getBoolean("tinactory.dependencyChecker.traceBlockedMethods")) {
             for (var method : methodsByOutput.getOrDefault(target, Set.of())) {
                 lines.add("  blocked method: " + method.id() + " (" + method.source() + ")");
                 method.requirements().stream()
-                    .filter(requirement -> !requirement.isSatisfied(this))
+                    .filter(requirement -> !requirement.isSatisfied(solver))
                     .map(requirement -> "    missing requirement: " + requirement.displayId())
                     .forEach(lines::add);
             }
@@ -823,9 +652,5 @@ public final class DependencyChecker implements IDependencyChecker {
     private static Collection<? extends IRecipe<?>> allTinyRecipes(IRecipeManager recipeManager,
         IRecipeType<?> recipeType) {
         return recipeManager.getAllRecipesFor((IRecipeType) recipeType);
-    }
-
-    private static DependencyMethod bootstrapMethod() {
-        return new DependencyMethod("bootstrap/start", List.of(), List.of(), "start list bootstrap");
     }
 }
